SWMRD-19 | Implement usage example notebook and summary method; Refactor as a package; Minor big fix;
