--- LC/data_containers.py
@@ -1,143 +1,0 @@
-# coding=utf-8
-
-from dataclasses import dataclass, asdict, fields, make_dataclass
-import numpy as np
-from typing import Tuple, Type
-from abc import ABC
-from collections import namedtuple
-
-
-@dataclass()
-class AbstractFeatureDataclass():
-    """ An abstract base dataclass for control related data manipulation
-
-    - (nice to have) ToDo: Add option to use timestamp instead of timestep index
-    """
-    feature_name: str
-    timestep_index: np.ndarray
-
-    @classmethod
-    def get_dimension_names(cls) -> Tuple[str, ...]:
-        container_properties = fields(cls)[2:]  # Remove the 'feature_name' and `timestep_index` properties
-        return tuple(each_field.name for each_field in container_properties)
-
-    @property
-    def shape(self) -> tuple[int, int]:
-        return self.__getattribute__(self.get_dimension_names()[0]).shape
-
-    @property
-    def trajectory_len(self) -> int:
-        return len(self.timestep_index)
-
-    @property
-    def sample_size(self) -> int:
-        return self.__getattribute__(self.get_dimension_names()[0]).size / self.trajectory_len
-
-    def ravel_dimensions_in_place(self) -> None:
-        for each_data_property in self.get_dimension_names():
-            ravel__copy = self.__getattribute__(each_data_property).ravel()
-            self.__setattr__(each_data_property, ravel__copy)
-        return None
-
-    def post_init_callback(self):
-        """ Overide methode to execute custom computation on feature dataclass """
-        pass
-
-    def __post_init__(self):
-
-        if not self.get_dimension_names():
-            raise TypeError(
-                    f"(!) AbstractFeatureDataclass is an abstract baseclass, "
-                    f"it must be subclassed in order to be instanciated.")
-
-        data_property_shape = None
-        for each_name in self.get_dimension_names():
-            data_property: np.ndarray = self.__getattribute__(each_name)
-
-            if type(data_property) is not np.ndarray:
-                raise TypeError(f"(!) Property `{each_name}` is not a numpy ndarray")
-            elif data_property_shape is not None:
-                if data_property_shape != data_property.shape:
-                    raise ValueError(f"(!) `{self.feature_name}` container received numpy array's with uneven shape")
-
-                data_property_shape = data_property.shape
-            else:
-                data_property_shape = data_property.shape
-
-            self.post_init_callback()
-
-        return None
-
-
-FeatureSpecification = namedtuple('FeatureSpecification', ['feature_dataclass_type', 'dimension_names'])
-
-
-def feature_dataclass_factory(specification: dict[str, list[str, ...]]) -> Type[AbstractFeatureDataclass]:
-    """ Factory to dynnamicaly create new `AbstractFeatureDataclass` subclass from dictionary specification.
-
-    Usage example:
-
-    >>> spec_ = dict(feature_dataclass_type='my_new_feature_type', dimension_names=['xx', 'yy', 'yawww'])
-    >>> the_new_cls = feature_dataclass_factory(specification=spec_)
-    >>> assert issubclass(the_new_cls, AbstractFeatureDataclass)
-    >>> # True
-    >>> assert isinstance(the_new_cls, AbstractFeatureDataclass)
-    >>> # False
-
-    :param specification: A dictionary with keys:value 'feature_dataclass_type': str, 'dimension_names': list[str, ...]}
-    :return: A new subclass of AbstractFeatureDataclass
-    """
-    try:
-        feature_dataclass_type: str = specification['feature_dataclass_type']
-        dimension_names: Tuple[str, ...] = specification['dimension_names']
-        if (type(feature_dataclass_type) is not str) or (type(dimension_names) is not list) or (
-                any([type(name) is not str for name in dimension_names])):
-            raise AttributeError
-    except KeyError as e:
-        raise KeyError(
-                f"(!) The key {e} is missing from the `specification` dictionary passed in argument.")
-    except AttributeError as e:
-        raise AttributeError(
-                f"(!) The `specification` dictionary must contain key 'feature_dataclass_type' with value of type "
-                f"string and key `dimension_names` with value of type list of string. {e}")
-
-    properties_field = [(name, np.ndarray) for name in dimension_names]
-
-    return make_dataclass(feature_dataclass_type, bases=(AbstractFeatureDataclass,), fields=properties_field)
-
-
-@dataclass()
-class StatePose2D(AbstractFeatureDataclass):
-    x: np.ndarray
-    y: np.ndarray
-    yaw: np.ndarray
-
-
-@dataclass()
-class CmdStandard(AbstractFeatureDataclass):
-    linear_vel: np.ndarray
-    angular_vel: np.ndarray
-
-
-@dataclass()
-class CmdSkidSteer(AbstractFeatureDataclass):
-    left: np.ndarray
-    right: np.ndarray
-
-
-@dataclass()
-class Velocity(CmdStandard):
-    pass
-
-
-@dataclass()
-class VelocitySkidSteer(CmdSkidSteer):
-    pass
-
-
-@dataclass()
-class DisturbanceQueryState:
-    state_pose_k: StatePose2D
-    velocity_skid_steer_k_previous: VelocitySkidSteer
-    cmd_skid_steer_k: CmdSkidSteer
-    cmd_skid_steer_k_previous: CmdSkidSteer
