+++ LC/plot_fct.py
