+++ multifeature_aggregator/__init__.py
@@ -1,0 +1,4 @@
+# coding=utf-8
+
+from .aggregator import aggregate_multiple_features
+from .data_containers import StatePose2D, CmdStandard, CmdSkidSteer, Velocity, VelocitySkidSteer
